[TS] Implement TextEncoder/TextDecoder with native code (#3800)
# Description of Changes

This implements (a subset of) the TextEncoder/TextDecoder web APIs using
native functions to do the actual `Uint8Array <-> String` conversion.
This should be a good bit faster than the `fast-text-encoding` package.

# Expected complexity level and risk

2 - this introduces new kinds of JS code and host calls to the v8 host,
but they're pretty well encapsulated.

# Testing

- [x] All TS modules make use of these already for encoding/decoding
strings in BSATN - the `fast-text-encoding` polyfill we pull in only
takes effect if the classes don't already exist, so the smoketests
passing means it works.