Handle unknown transactions without erroring. (#3858)
# Description of Changes

Don't throw an error if there is no `reducerInfo`. The code was
previously trying to handle the case of an unknown reducer, but was
effectively asserting that the reducerInfo existed too soon.

Now we should be fine handling a transaction, even if we can't determine
reducer information for it.

# Expected complexity level and risk

1.

# Testing

I tested this by running the quickstart chat app, and using the CLI to
delete rows (via `spacetime sql`). Before this change, the client
errors, but now it handles it correctly.

I also tested with the repro in
https://github.com/clockworklabs/SpacetimeDB/issues/3817