Add a fault-tolerant commitlog replay mode for use in debugging (#3887)
# Description of Changes

When debugging broken commitlogs, we want to inspect the whole
commitlog, including the part after the first error.
This is in contrast with the way we want to replay in prod, where we'd
rather get a hard error than an incorrect state.

This commit adds a new flag to commitlog replay, `ErrorBehavior`. The
`core` crate passes `ErrorBehavior::FailFast`
when replaying commitlogs to reconstruct databases. Internal tooling
(not in this repository) uses `ErrorBehavior::Warn` to print the
entirety of a broken commitlog.

# API and ABI breaking changes

Changes internal APIs only.

# Expected complexity level and risk

1 - no change to behavior of SpacetimeDB.

# Testing

None.