Resolve algebraic type refs recursively for view type check (#3876)
# Description of Changes

Resolves algebraic type refs recursively in order to check the product
type of a query builder view.

This should fix the issue reported
[here](https://discord.com/channels/1037340874172014652/1448796556366057513).
However I've so far been unsuccessful in trying repro it.

Also adds further commentary to `Typespace::resolve` to make it clear
that it is not recursive.

# API and ABI breaking changes

None

# Expected complexity level and risk

0

# Testing

TODO. So far I haven't been able to repro with a smoketest